+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # RELEASE NOTES
 
+## Next
+* DNS
+    * Fail on attempted Zone deletion. Not supported.
+    * Create SOA and NS Records on zone read if don't exist.
+    * Add HTTPS, SVCB record support
+* GTM
+    * Add validation for property type and traffic targets combination
+
+## 1.2.1 (Feb 3, 2021)
+
+#### BUG FIXES:
+* PAPI -- Fixed crash caused by passing computed cpCode as a variable in rules to akamai_property
+* PAPI -- Deprecated "product" attribute in akamai_cp_code resource and changed it "product_id"
+
 ## 1.2.0 (Jan 14, 2021) Identity and Access Management support
 
 These are the operations supported in the Identity Management: User Administration API v2: